+++ lib/statistics/providers.rb
@@ -2,3 +2,5 @@ module Statistics
   module Providers
   end
 end
+
+require_relative 'providers/connpass'
